% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_project}
\alias{create_project}
\title{create essential project folder}
\usage{
create_project(name = NULL, path = NULL)
}
\arguments{
\item{name}{project name}

\item{path}{directory of the project}
}
\description{
create essential project folder
}
