% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_info}
\alias{file_info}
\title{list all files with detailed information in a folder}
\usage{
file_info(
  path = ".",
  pattern = NULL,
  recursive = TRUE,
  full.names = TRUE,
  ignore.case = FALSE,
  all.files = FALSE
)
}
\arguments{
\item{path}{Full path to list.}

\item{pattern}{Match patter. See \code{\link{list.files}}.}

\item{recursive}{Recursive into sub-folders.}

\item{full.names}{List full name.}

\item{ignore.case}{Case-insensitive.}

\item{all.files}{Whether to show invisible files.}
}
\description{
List all files with detailed information in a folder. This function is very useful to check large files, eg. GWAS summary statistics.
}
