% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggsave_all}
\alias{ggsave_all}
\title{ggsave in different formats}
\usage{
ggsave_all(filename, ...)
}
\arguments{
\item{filename}{Filename}

\item{\dots}{Further arguments to pass.}
}
\description{
ggsave in different formats, including pdf, tiff, eps, and png (different journals may require different formats).
}
