% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{meta_inverse_variance}
\alias{meta_inverse_variance}
\alias{meta_inverse_variance_df}
\title{meta-analysis based on inverse variance method.}
\usage{
meta_inverse_variance(beta1, beta2, se1, se2)

meta_inverse_variance_df(data = NULL, beta1, beta2, se1, se2)
}
\arguments{
\item{beta1}{effect size for the first study.}

\item{beta2}{effect size for the second study.}

\item{se1}{standard error for the first study.}

\item{se2}{standard error for the second study.}

\item{data}{if a data.table, the beta and se are the column names.}
}
\description{
meta-analysis based on inverse variance method for a data.
}
