% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_venn}
\alias{plot_venn}
\title{venn plot}
\usage{
plot_venn(
  x,
  color = NULL,
  category.names = NULL,
  filename = NULL,
  imagetype = "png",
  height = 800,
  width = 1200,
  units = "px",
  resolution = 300,
  compression = "lzw",
  lwd = 3,
  lty = "blank",
  fill = color,
  cex = 0.8,
  fontfamily = "sans",
  cat.cex = 0.8,
  cat.default.pos = "outer",
  cat.pos = c(0, 0, 0, 0, 0, 0, 0),
  cat.dist = c(0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1)
)
}
\arguments{
\item{x}{A list of data.}

\item{color}{color for the venn, default is Set1 from brewer.pal.}

\item{category.names}{label, default is the names in the list.}

\item{filename}{file name to save}

\item{imagetype}{"png" or "tiff"}

\item{height}{default value 800}

\item{width}{default value 1200}

\item{units}{default value "px",}

\item{resolution}{default value 300}

\item{compression}{default value "lzw"}

\item{lwd}{default value 3,}

\item{lty}{default value 'blank',}

\item{fill}{default value is the same as color}

\item{cex}{default value .8,}

\item{fontfamily}{default value "sans",}

\item{cat.cex}{default value 0.8,}

\item{cat.default.pos}{default value "outer",}

\item{cat.pos}{default value c(0, 0, 0, 0, 0, 0, 0),}

\item{cat.dist}{default value c(0.1, 0.1, 0.1, 0.1, 0.1,0.1, 0.1)}
}
\description{
A wrapper function of venn plot.
save_plot in different formats, including pdf, tiff, eps, and png (different journals may require different formats).
}
