% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_last_line}
\alias{read_last_line}
\title{read the last line of all files in a folder}
\usage{
read_last_line(
  path = ".",
  pattern = NULL,
  recursive = TRUE,
  full.names = TRUE,
  ignore.case = FALSE,
  all.files = FALSE
)
}
\arguments{
\item{path}{Full path to list.}

\item{pattern}{Match patter. See \code{\link{list.files}}.}

\item{recursive}{Recursive into sub-folders.}

\item{full.names}{List full name.}

\item{ignore.case}{Case-insensitive.}

\item{all.files}{Whether to show invisible files.}
}
\description{
Read the last line of all files in a folder. This function is useful to check log files.
}
