% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{remove_duplicate_NA}
\alias{remove_duplicate_NA}
\title{remove duplicates and NA from a vector}
\usage{
remove_duplicate_NA(x)
}
\arguments{
\item{x}{A vector}
}
\description{
remove duplicates and NA from a vector
}
