% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot_all}
\alias{save_plot_all}
\title{save_plot in different formats}
\usage{
save_plot_all(filename, ...)
}
\arguments{
\item{filename}{Filename}

\item{\dots}{Further arguments to pass.}
}
\description{
save_plot in different formats, including pdf, tiff, eps, and png (different journals may require different formats).
}
